disable deletion

+++ src/main/java/com/nordstrom/kafka/connect/sqs/SqsSourceConnectorTask.java
@@ -1,12 +1,12 @@
 /*
  * Copyright 2019 Nordstrom, Inc.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
  * the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
@@ -39,7 +39,7 @@ public class SqsSourceConnectorTask extends SourceTask {
 
   /*
    * (non-Javadoc)
-   * 
+   *
    * @see org.apache.kafka.connect.connector.Task#version()
    */
   @Override
@@ -49,7 +49,7 @@ public class SqsSourceConnectorTask extends SourceTask {
 
   /*
    * (non-Javadoc)
-   * 
+   *
    * @see org.apache.kafka.connect.source.SourceTask#start(java.util.Map)
    */
   @Override
@@ -65,7 +65,7 @@ public class SqsSourceConnectorTask extends SourceTask {
 
   /*
    * (non-Javadoc)
-   * 
+   *
    * @see org.apache.kafka.connect.source.SourceTask#poll()
    */
   @Override
@@ -112,13 +112,16 @@ public class SqsSourceConnectorTask extends SourceTask {
     final String receipt = record.sourceOffset().get( SqsConnectorConfigKeys.SQS_MESSAGE_RECEIPT_HANDLE.getValue() )
         .toString() ;
     log.debug( ".commit-record:url={}, receipt-handle={}", config.getQueueUrl(), receipt ) ;
-    client.delete( config.getQueueUrl(), receipt ) ;
+    System.out.println("config.isDeletionEnable() = " + config.isDeletionEnable());
+    if (config.isDeletionEnable()) {
+     client.delete( config.getQueueUrl(), receipt ) ;
+    }
     super.commitRecord( record ) ;
   }
 
   /*
    * (non-Javadoc)
-   * 
+   *
    * @see org.apache.kafka.connect.source.SourceTask#stop()
    */
   @Override
@@ -129,7 +132,7 @@ public class SqsSourceConnectorTask extends SourceTask {
   /**
    * Test that we have both the task configuration and SQS client properly
    * initialized.
-   * 
+   *
    * @return true if task is in a valid state.
    */
   private boolean isValidState() {
